requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.7.0
sqlmodel>=0.0.18
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
anyio>=4.3.0
psycopg2-binary>=2.9.0
openai>=1.0.0
